# openapi.yaml
openapi: 3.1.0
info:
  title: POVIEW.AI Neo4j Memory Builder API
  version: 2.0.0
  description: >
    A secure API wrapper that connects to a Neo4j Aura instance to store and retrieve structured entities, relationships,
    and insights. Supports dynamic labels (Person, Company, Project, etc.), configurable write modes, safe query execution,
    and visualization-ready graph output for dashboards.

servers:
  - url: https://neo4j-memory-api-production.up.railway.app
    description: Deployed Neo4j Memory API server

paths:

  /write:
    post:
      summary: Write structured entity with dynamic labels and relationships
      description: >
        Creates or updates a node in Neo4j. Supports configurable modes (create, skip, overwrite) and optional
        relationships between nodes.
      operationId: write
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                text:
                  type: string
                  description: Primary text or name of the node.
                label:
                  type: string
                  default: Memory
                  description: >
                    Node label (e.g., Person, Company, Project, Conversation, Reply, Insight, Collateral, etc.)
                mode:
                  type: string
                  enum: [create, skip, overwrite]
                  default: create
                  description: |
                    Controls write behavior:
                    - `create`: Always create a new node
                    - `skip`: Only create if missing
                    - `overwrite`: Update if exists
                context:
                  type: object
                  description: Optional metadata to attach to the node.
                  example:
                    role: "CEO"
                    company: "Oakland Care Group"
                    sector: "Healthcare"
                relationships:
                  type: array
                  description: Optional relationships to create.
                  items:
                    type: object
                    properties:
                      from:
                        type: string
                        description: The `text` value of the source node.
                      to:
                        type: string
                        description: The `text` value of the target node.
                      type:
                        type: string
                        description: The relationship type (e.g., WORKS_AT, CREATED_FOR, INFORMED_BY).
              required: [text]
      responses:
        "200":
          description: Entity successfully written or updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  label:
                    type: string
                  mode:
                    type: string
                  nodeId:
                    type: string
                    description: ID of the created or matched node
        "400":
          description: Invalid input or missing parameters
        "500":
          description: Internal server error

  /query:
    post:
      summary: Execute Cypher queries or presets
      description: >
        Runs a Cypher query or built-in preset (e.g., `getGraphSnapshot`, `getAllPersons`, `getInsights`).
        Supports parameter substitution and multiple output formats.
      operationId: query
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                cypher:
                  type: string
                  description: Custom Cypher query to execute.
                  example: "MATCH (p:Person)-[:WORKS_AT]->(c:Company) RETURN p,c LIMIT 50"
                params:
                  type: object
                  description: Optional parameters for the Cypher query.
                  example:
                    limit: 100
                preset:
                  type: string
                  enum: [getAllPersons, getAllCompanies, getGraphSnapshot, getInsights]
                  description: Optional named preset query.
                format:
                  type: string
                  enum: [records, json]
                  default: json
                  description: Output format.
                limit:
                  type: integer
                  default: 250
                  description: Record limit for large queries.
      responses:
        "200":
          description: Query executed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  records:
                    type: integer
                    description: Number of records returned
                  format:
                    type: string
                    example: json
                  preset:
                    type: string
                    example: getGraphSnapshot
                  results:
                    type: array
                    items:
                      type: object
                      description: Cypher record or parsed JSON result
        "400":
          description: Invalid or unsafe query
        "500":
          description: Database or query error

  /graph:
    post:
      summary: Return visualization-ready graph data
      description: >
        Returns compact graph data (`nodes`, `links`) suitable for dashboards and visual network tools like PyVis, D3, or Plotly.
      operationId: graph
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                limit:
                  type: integer
                  default: 500
                  description: Maximum number of relationships to return.
                filterLabel:
                  type: string
                  description: Optional node label filter (e.g., "Person", "Company").
      responses:
        "200":
          description: Graph snapshot returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  nodes:
                    type: array
                    description: Node array for visualization
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        label:
                          type: string
                        text:
                          type: string
                        context:
                          type: object
                  links:
                    type: array
                    description: Relationship array for visualization
                    items:
                      type: object
                      properties:
                        source:
                          type: string
                        target:
                          type: string
                        type:
                          type: string
        "500":
          description: Graph query error

  /health:
    get:
      summary: Health check
      description: Confirms database connectivity and service status.
      operationId: health
      responses:
        "200":
          description: Service operational
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok